[DDD] Fixed naming to hide implementation details
