--- src/lib/Storage/MetricsInterface.php
@@ -1,14 +1,0 @@
-<?php
-
-/**
- * @copyright Copyright (C) eZ Systems AS. All rights reserved.
- * @license For full copyright and license information view LICENSE file distributed with this source code.
- */
-declare(strict_types=1);
-
-namespace EzSystems\EzSupportTools\Storage;
-
-interface MetricsInterface
-{
-    public function getValue(): int;
-}
