+++ src/lib/Storage/MetricsProvider.php
@@ -8,26 +8,7 @@ declare(strict_types=1);
 
 namespace EzSystems\EzSupportTools\Storage;
 
-use EzSystems\EzSupportToolsBundle\SystemInfo\Exception\MetricsNotFoundException;
-
-final class MetricsProvider implements MetricsProviderInterface
+interface MetricsProvider
 {
-    /** @var iterable */
-    private $metrics;
-
-    public function __construct(iterable $metrics)
-    {
-        $this->metrics = $metrics;
-    }
-
-    public function provideMetrics(string $identifier): MetricsInterface
-    {
-        foreach ($this->metrics as $metricKey => $metric) {
-            if ($metricKey === $identifier) {
-                return $metric;
-            }
-        }
-
-        throw new MetricsNotFoundException($identifier);
-    }
+    public function provideMetrics(string $identifier): Metrics;
 }
